+++ field/field-object.js
@@ -8,6 +8,7 @@ const ErrorFieldNotAllowed = require('error-types').ErrorFieldNotAllowed;
 const FieldText = require('./field-text').FieldText;
 const FieldGuid = require('./field-text').FieldTextGuid;
 const FieldTextBoolean = require('./field-text-boolean').FieldTextBoolean;
+const FieldTextBinary = require('./field-text-binary').FieldTextBinary;
 
 class FieldObject extends Field {
 
@@ -28,7 +29,8 @@ class FieldObject extends Field {
       typeGuid: new FieldGuid({emptyAllow:  this.emptyAllow }),      // the id, overrules the type
       typeIsDefault: new FieldGuid({emptyAllow: this.emptyAllow}),   // set in the code table it's default
       typeInsertOnly: new FieldTextBoolean({emptyValueAllowed: this.emptyValueAllowed}),
-                                                                     // if set to true the code will not be delete by the sync
+      _mode: new FieldTextBinary({emptyAllow: this.emptyValueAllowed}),
+      // if set to true the code will not be delete by the sync
 
       fieldTypeId:  new FieldGuid({emptyAllow:  this.emptyAllow }),  // if set: set the code.typeId
       fieldTypeGuid: new FieldGuid({emptyAllow:  this.emptyAllow }), // or find the code.guid => typeId