+++ test/field-text-binary.spec.js
@@ -1,0 +1,12 @@
+const Chai = require('chai');
+const assert = Chai.assert;
+const _ = require('lodash');
+const Logger = require('logger');
+const Lookup = require('../lib/lookup');
+const FieldTextBinary = require('../field/field-text-binary').FieldTextBinary;
+
+describe('field.text.binary', () => {
+
+  let logger = new Logger({toConsole: false});
+  // write the tests
+});